Move uncompress.c file
